requirements.txt
appdirs==1.4.3
certifi==2018.1.18
Cython==0.23
Django==1.10.6
docutils==0.13.1
packaging==16.8
pipenv==11.8.0
psutil==5.0.1
pyowm
pytz
pandas
datetime
Pillow
pillow==7.1.2 
Pygments==2.2.0
pyparsing==2.2.0
aiogram
aiohttp
asyncio
requests==2.13.0
six==1.10.0
virtualenv==15.1.0
virtualenv-clone==0.3.0